+++ ORBITER/src/lib/group_actions/actions/actions.h
@@ -1096,12 +1096,12 @@ public:
 		int *Adj, int n, int *labeling, int verbose_level);
 	// labeling[n]
 	action *create_automorphism_group_of_block_system(
-		int nb_points, int nb_blocks, int block_size, int *Blocks,
+		int nb_points, int nb_blocks, int block_size, long int *Blocks,
 		int verbose_level);
 	action *create_automorphism_group_of_collection_of_two_block_systems(
 		int nb_points,
-		int nb_blocks1, int block_size1, int *Blocks1,
-		int nb_blocks2, int block_size2, int *Blocks2,
+		int nb_blocks1, int block_size1, long int *Blocks1,
+		int nb_blocks2, int block_size2, long int *Blocks2,
 		int verbose_level);
 	action *create_automorphism_group_of_incidence_matrix(
 		int m, int n, int *Mtx,
